+++ Sources/FlipView/_Deprecated/View+Deprecated.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+
+public extension View {
+
+    @available(*, deprecated, message: "This is no longer needed, and can be removed.")
+    func withFlipViewListBugFix() -> some View {
+        self
+    }
+}